Fix incorrect matrix_row_t size problem.
The matrix.h define the size of matrix_row_t, but it's not check whether the MATRIX_COLS is correct.
This patch would include the missing config.h file, make the size correct.

The incorrect size of matrix_row_t would cause key at the colomn which greater than 8 missing.